requirements.txt
langchain
langchain-core
langchain-openai
openai
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn
pandas
seaborn
matplotlib
langchain-community
pypdf
requests
beautifulsoup4
